Moves Logger classes from storage to logger package Creates header on file creation
